pyproject.toml
[tool.poetry]
name = "test-ui"
version = "0.1.0"
description = ""
authors = ["Mikhail Chuvikin"]
readme = "README.md"
packages = [{include = "test_ui"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.4.0"
allure-pytest = "^2.13.2"
requests = "^2.31.0"
curlify2 = "^2.0.0"
dynaconf = "^3.2.2"
selenium = "^4.12.0"
pytest-playwright = "^0.4.2"
pytest-assume = "^2.4.3"
pydantic = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
